fix crash from types.
https://sponge.corp.google.com/target?id=91b5fd06-360c-465d-b3ec-40a6ee9a9c98&target=earth_vr%2Fcontinuous_test_prod_x64&searchFor=

PiperOrigin-RevId: 162643823

+++ src/utilities.cpp
@@ -279,8 +279,8 @@ bool SavePreference(const char *key, int32_t value) {
 }
 
 #if defined(_WIN32)
-inline char *getcwd(char *buffer, int maxlen) {
-  return _getcwd(buffer, maxlen);
+inline char *getcwd(char *buffer, size_t maxlen) {
+  return _getcwd(buffer, static_cast<int>(maxlen));
 }
 
 inline int chdir(const char *dirname) { return _chdir(dirname); }
